Added logic for adding remaining consultants
